move choiceworld.py to pose_dlc.py (backwards compatible)
